# openapi_v0.8.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "sisRUA API",
    "description": "\n**sisRUA** - Generative Urban Design System for AutoCAD.\n\nThis API powers the AutoCAD plugin for:\n- Fetching and projecting **OpenStreetMap** data\n- Processing **GeoJSON** files for CAD import\n- Providing **elevation data** from SRTM sources\n- Managing **asynchronous jobs** for long-running operations\n\n## Authentication\nProtected endpoints require the `X-SisRua-Token` header.\n    ",
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/audit": {
      "post": {
        "tags": [
          "Audit"
        ],
        "summary": "Create Audit Log",
        "description": "Create audit log entry (called from C# plugin or other services).",
        "operationId": "create_audit_log_api_audit_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "List Audit Logs",
        "description": "List audit logs with optional filters.",
        "operationId": "list_audit_logs_api_audit_get",
        "parameters": [
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Type"
            }
          },
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Type"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/{audit_id}": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Get Audit Log",
        "description": "Get a specific audit log entry.",
        "operationId": "get_audit_log_api_audit__audit_id__get",
        "parameters": [
          {
            "name": "audit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/{audit_id}/verify": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "Verify Audit Log",
        "description": "Verify audit log signature to detect tampering.",
        "operationId": "verify_audit_log_api_audit__audit_id__verify_get",
        "parameters": [
          {
            "name": "audit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Audit Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}